legacy/python/requirements.txt
# Core
neo4j>=5.0.0
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0

# Optional: Go support
tree-sitter-go>=0.20.0

# Optional: Git history
gitpython>=3.1.0

# Optional: Docker observation
docker>=6.0.0

# Embeddings
numpy>=1.24.0
sentence-transformers>=2.2.0

# Vector persistence
chromadb>=0.4.0
